UPDATED: Created various fixes for various problems, some to do with making a 64-bit build of the library, all of which are included in this commit

+++ libHTTP/init.c
@@ -5,7 +5,7 @@
 #include "internals.h"
 
 BOOL Initialised = FALSE;
-int httpError = 0;
+long httpError = 0;
 
 /** Initialises the library
  * @return error state of the call, 0 for success, -1 for already initialised, and 1000+ for a Windows Sockets Error
